fix(connector): support cache durability on Windows
